requirements.txt
setuptools~=58.0.4
numpy~=1.21.2
websocket-client~=1.2.1
msgpack~=1.0.2